Extract english locale from distanceInWords
